Added delete organization if you are siteadmin

+++ src/app/backend-api/identity-registry/services/organizations.service.ts
@@ -21,9 +21,12 @@ export class OrganizationsService implements OnInit {
 	}
 
 	public deleteOrganization(orgMrn:string):Observable<any> {
-		// TODO: for now this organization is crucial to the Maritime Cloud, and cannot be deleted
+		// TODO: for now this organization is crucial to the Maritime Cloud, and cannot be deleted. Also you cant delete you own
 		if (orgMrn === 'urn:mrn:mcl:org:dma') {
 			throw Error('You cannot delete this organization');
+		}
+		if (this.authService.isMyOrg(orgMrn)) {
+			throw Error('You cannot delete your own organization');
 		}
 		return this.organizationApi.deleteOrgUsingDELETE(orgMrn);
 	}
